requirements.txt
pandas
numpy
scikit-learn
xgboost
mlflow
boto3
fastapi
uvicorn
python-dotenv
kaggle
shap
